+++ tests/todo-diagnostic-report.test.mjs
@@ -1,9 +1,0 @@
-import test from 'node:test';
-
-test.todo(
-  'add getDiagnosticReportClaimsFromCommunicationDocument(...) using shared DiagnosticReport claim keys from gdc-common-utils-ts',
-);
-
-test.todo(
-  'cover DiagnosticReport.contained-documents and presented-form-* extraction from Communication bundle payloads',
-);